Renamed UserDetails module
