+++ components/UserDetail/Stats.tsx
